refactor(llment): organize commands into modules
