+++ crates/llment/src/commands/mod.rs
@@ -1,0 +1,13 @@
+pub mod clear;
+pub mod model;
+pub mod prompt;
+pub mod provider;
+pub mod quit;
+pub mod redo;
+
+pub use clear::ClearCommand;
+pub use model::ModelCommand;
+pub use prompt::PromptCommand;
+pub use provider::ProviderCommand;
+pub use quit::QuitCommand;
+pub use redo::RedoCommand;
